requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
google-generativeai==0.3.0
# Google Gemini API
google-generativeai>=0.3.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
plotly==5.18.0
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
lxml==4.9.3
requests==2.31.0
pydantic==2.5.0
loguru==0.7.2
pytest==7.4.3
psycopg2-binary==2.9.7  # optional: run scripts/apply_migrations.py
